Use richtext for chatmessages. This change enables us to format the messages for easier differentiation between the players, system messages and so on. Currently normal chat messages begin with underlined and colored playername, /me messages are completely colored and italic style and system messages are bold text. Spectators (as well as system messages) use grey color
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3725 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/wlapplication.cc
@@ -1449,7 +1449,7 @@ struct SinglePlayerGameSettingsProvider : public GameSettingsProvider {
 	}
 
 	virtual void setPlayerNumber(int32_t number) {
-		if (number >= s.players.size())
+		if (number >= static_cast<int32_t>(s.players.size()))
 			return;
 
 		s.playernum = number;
